ta: fix compiler warnings related to GP compat updates
With recent update, some arguments changed from uint32_t to size_t. This
made the TA panic as well. Change to size_t accordingly.

+++ locki/ta/include/ta_locki_crypto.h
@@ -48,6 +48,6 @@ TEE_Result sha256(const uint8_t *in, const size_t inlen, uint8_t *digest);
 
 TEE_Result hmac_sha256(const uint8_t *key, const size_t keylen,
 		       const uint8_t *in, const size_t inlen,
-		       uint8_t *out, uint32_t *outlen);
+		       uint8_t *out, size_t *outlen);
 
 #endif
